requirements.txt
accelerate==1.12.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
albucore==0.0.24
albumentations==2.0.8
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
astor==0.8.1
async-timeout==4.0.3
attrs==25.4.0
av==16.1.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.14.3
blake3==1.0.8
boto3==1.42.42
botocore==1.42.42
brotli==1.2.0
build==1.4.0
cachetools==7.0.0
cbor2==5.8.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
chromadb==1.4.1
click==8.3.1
cloudpickle==3.1.2
coloredlogs==15.0.1
colorlog==6.10.1
compressed-tensors==0.11.0
contourpy==1.3.2
cryptography==46.0.4
cupy-cuda12x==13.6.0
cycler==0.12.1
dataclasses-json==0.6.7
depyf==0.19.0
dill==0.4.1
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
doclayout_yolo==0.0.4
durationpy==0.10
einops==0.8.2
email-validator==2.3.0
exceptiongroup==1.3.1
fast-langdetect==0.2.5
fastapi==0.128.1
fastapi-cli==0.0.20
fastapi-cloud-cli==0.11.0
fastar==0.8.0
fastrlock==0.8.3
fasttext-predict==0.9.2.4
ffmpy==1.0.0
filelock==3.20.3
flatbuffers==25.12.19
fonttools==4.61.1
frozendict==2.4.7
frozenlist==1.8.0
fsspec==2026.1.0
ftfy==6.3.1
gguf==0.17.1
googleapis-common-protos==1.72.0
gradio==5.49.1
gradio_client==1.13.3
gradio_pdf==0.0.22
greenlet==3.3.1
groovy==0.1.2
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-retries==0.4.5
httpx-sse==0.4.3
huggingface_hub==0.36.1
humanfriendly==10.0
idna==3.11
ImageIO==2.37.2
importlib_metadata==8.7.1
importlib_resources==6.5.2
interegular==0.3.3
Jinja2==3.1.6
jiter==0.13.0
jmespath==1.1.0
joblib==1.5.3
json_repair==0.56.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
kubernetes==35.0.0
langchain==1.2.8
langchain-chroma==1.1.0
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.9
langchain-huggingface==1.2.0
langchain-openai==1.1.7
langchain-tavily==0.2.17
langchain-text-splitters==1.1.0
langchainhub==0.1.21
langgraph==1.0.8
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.4
langsmith==0.6.9
lark==1.2.2
lazy_loader==0.4
llguidance==0.7.30
llvmlite==0.44.0
lm-format-enforcer==0.11.3
loguru==0.7.3
magic-pdf==1.3.12
magika==1.0.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.8
mdurl==0.1.2
mineru==2.7.5
mineru_vl_utils==0.1.22
mistral_common==1.9.0
mmh3==5.2.0
modelscope==1.34.0
mpmath==1.3.0
msgpack==1.1.2
msgspec==0.20.0
multidict==6.7.1
mypy_extensions==1.1.0
networkx==3.4.2
ninja==1.13.0
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
oauthlib==3.3.1
omegaconf==2.3.0
onnxruntime==1.23.2
openai==1.109.1
openai-harmony==0.0.8
opencv-python==4.13.0.90
opencv-python-headless==4.13.0.90
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.7
ormsgpack==1.12.2
outlines_core==0.2.11
overrides==7.7.0
packaging==24.2
pandas==2.3.3
partial-json-parser==0.2.1.1.post7
pdfminer.six==20260107
pdftext==0.6.3
pillow==11.3.0
polars==1.38.0
polars-runtime-32==1.38.0
posthog==5.4.0
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.24.1
propcache==0.4.1
protobuf==6.33.5
psutil==7.2.2
py-cpuinfo==9.0.0
pybase64==1.4.3
pyclipper==1.4.0
pycountry==24.6.1
pycparser==3.0
pydantic==2.12.5
pydantic-extra-types==2.11.0
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
PyMuPDF==1.24.14
pyparsing==3.3.2
pypdf==6.6.2
pypdfium2==4.30.0
PyPika==0.51.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-multipart==0.0.22
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
qwen-vl-utils==0.0.14
rapid-table==1.0.5
ray==2.53.0
referencing==0.37.0
regex==2026.1.15
reportlab==4.4.9
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.3.2
rich-toolkit==0.18.1
rignore==0.7.6
robust-downloader==0.0.2
rpds-py==0.30.0
ruff==0.15.0
s3transfer==0.16.0
safehttpx==0.1.7
safetensors==0.7.0
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
semantic-version==2.10.0
sentence-transformers==5.2.2
sentencepiece==0.2.1
sentry-sdk==2.52.0
setproctitle==1.3.7
shapely==2.1.2
shellingham==1.5.4
simsimd==6.5.12
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.8.3
soxr==1.0.0
SQLAlchemy==2.0.46
starlette==0.50.0
stringzilla==4.6.0
sympy==1.14.0
tenacity==9.1.3
thop==0.1.1.post2209072238
threadpoolctl==3.6.0
tifffile==2025.5.10
tiktoken==0.12.0
tokenizers==0.22.2
tomli==2.4.0
tomlkit==0.13.3
torch==2.8.0
torchaudio==2.8.0
torchvision==0.23.0
tqdm==4.67.3
transformers==4.57.6
triton==3.4.0
typer==0.21.1
typer-slim==0.21.1
types-requests==2.32.4.20260107
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ultralytics==8.4.11
ultralytics-thop==2.0.18
urllib3==2.6.3
uuid_utils==0.14.0
uv==0.9.30
uvicorn==0.40.0
uvloop==0.22.1
vllm==0.11.0
watchfiles==1.1.1
wcwidth==0.5.3
websocket-client==1.9.0
websockets==15.0.1
xformers==0.0.32.post1
xgrammar==0.1.25
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
